refactor: Header 안쪽 components 폴더 제거 #439
